Clarify the game rules (#124)
* Clarify the game rules

This file is a much better place for all this information.

* Update PLAYING.md

Does this work?

* Update PLAYING.md

* Update PLAYING.md

* Update PLAYING.md

* Update PLAYING.md

* Update PLAYING.md

* Update PLAYING.md

* Update PLAYING.md

* Update PLAYING.md

Then, better to get rid of it

* Update PLAYING.md

* Clean ` vs '

I tried to clean apostrophes.

* Update PLAYING.md

* Update PLAYING.md

I have corrected the text a bit more:
`Pointsman` to `pointman`
`travel` to `cycle`
`stash` to `inventory`, but not everywhere. After all, it may be also confusing
Lines 548-550

* Update PLAYING.md

The limit of symbols per line was exceeded.

* Update PLAYING.md

I removed the `TAB` sentence.

* Update PLAYING.md

I have added most of the suggestions. However, I am unclear about what to change in lines 608-618.

* Update PLAYING.md

I think it would be more reasonable to keep the crafting section in the main help. I have cut a few lines that were definitely redundant.
But I don't mind if crafting moves to the Q&A section.

* Update PLAYING.md

During the previous attempt, I failed to incorporate changes. Let me try once again.

* Update PLAYING.md

* Update PLAYING.md

* Update PLAYING.md

* Update PLAYING.md

Reshaping lines 611 and 612

* Update GameDefinition/PLAYING.md

* Update GameDefinition/PLAYING.md

* Update PLAYING.md

Remove white spaces from the ends of all of the lines

---------

Co-authored-by: Mikolaj Konarski <mikolaj.konarski@gmail.com>